requirements.txt
bleach==2.1.3
certifi==2018.1.18
chardet==3.0.4
cycler==0.10.0
decorator==4.2.1
enum34==1.1.6
html5lib==1.0.1
idna==2.6
Markdown==2.6.11
networkx==2.1
numpy==1.14.2
Pillow==5.0.0
pkginfo==1.4.2
protobuf==3.5.2
pyparsing==2.2.0
python-dateutil==2.7.0
pytz==2018.3
PyWavelets==0.5.2
requests==2.18.4
requests-toolbelt==0.8.0
scikit-image==0.13.1
scipy==1.0.0
six==1.11.0
tensorflow-gpu==1.5.0
tensorflow-tensorboard==1.5.1
tensorlayer==1.7.3
tqdm==4.19.7
twine==1.10.0
urllib3==1.22
Werkzeug==0.14.1
coverage
easydict
logbook